Refactor Code
Moved individual folders to oldFiles. Moved login files into the root
directory so that _SESSSION can be used. "pageFiles" folder contains
the html code for the header, navbar, and footer. "views" folder
contains the different html views that will then be displayed between
the navbar and footer. "model" folder contains the php functions that
are called from the index.php file and contains the sql statements.
Again, the pdo_connect file is set to be ignored so that it isn't
uploaded to the repo and doesn't need to be worried about since it's
already setup on the server. I can send the file to you if you want to
look at the code though.